integration of auth system

+++ src/routers/tasksRouter.js
@@ -25,19 +25,21 @@ tasksRouter.get('/', validations.authValidation(), routeHelper( authHelper(
 )));
 
 //Obtener una tarea
-tasksRouter.get('/:id', routeHelper( async (req, res) => {
-    let sqlQuery = 'SELECT id, title, description, date, userId FROM tasks where tasks.id=?';
-    const {id} = req.authData;
-    let rows = await pool.query(sqlQuery, id);
-    if (rows.length > 0) {
-        res.status(200).json(rows);
-    } else {
-        res.status(404).json({
-            status: 'error',
-            message: `task don't exist`,
-        });
+tasksRouter.get('/:id',validations.authValidation(), routeHelper( authHelper(
+    async (req, res) => {
+        let sqlQuery = 'SELECT id, title, description, date, userId FROM tasks where tasks.id=?';
+        const {id} = req.authData;
+        let rows = await pool.query(sqlQuery, req.params.id);
+        if (rows.length > 0 && rows[0].userId === id) {
+            res.status(200).json(rows);
+        } else {
+            res.status(404).json({
+                status: 'error',
+                message: `task don't exist or is not yours`,
+            });
+        }
     }
-}));
+)));
 
 
 //Crear una tarea
@@ -60,32 +62,48 @@ tasksRouter.post('/', validations.authValidation(), validations.validate(validat
 )));
 
 //Actualizar una tarea
-tasksRouter.put('/:id', validations.validate(validations.taskValidation), routeHelper( async (req, res) => {
-    const {title, date, description} = req.body;
-    let sqlQuery = 'UPDATE tasks SET title=?, date=?, description=? WHERE id=?';
-    let rows = await pool.query(sqlQuery, [title, date, description, req.params.id]);
-    if (rows.length > 0) {
-        res.status(200).json(rows[0]);
-    } else {
-            res.status(404).json({
-            status: 'error',
-            message: `task don't exist`,
-        });
+tasksRouter.put('/:id', validations.authValidation(), validations.validate(validations.taskValidation), routeHelper( authHelper(
+    async (req, res) => {
+        const {title, date, description} = req.body;
+        let sqlQuery = 'UPDATE tasks SET title=?, date=?, description=? WHERE id=? AND userId=?';
+        const {id} = req.authData
+        const dateFormatted = getSqlDate(Date(date))
+        let rows = await pool.query(sqlQuery, [title, dateFormatted, description, req.params.id, id]);
+        if (rows.affectedRows > 0) {
+            res.status(200).json({
+                status: 'ok',
+                updated: true,
+            });
+        } else {
+                res.status(404).json({
+                status: 'error',
+                message: `task don't exist or is not yours`,
+            });
+        }
     }
-}));
+)));
 
 //Eliminar una tarea
-tasksRouter.delete('/:id', routeHelper( async (req, res) => {
-    let sqlQuery = 'DELETE FROM tasks WHERE tasks.id=?';
-    let rows = await pool.query(sqlQuery, req.params.id);
-    if (rows.length > 0) {
-        res.status(200).json(rows[0]);
-    } else {
-        res.status(404).json({
-            status: 'error',
-            message: `task don't exist`,
-        });
+tasksRouter.delete('/:id', validations.authValidation(), routeHelper( authHelper(
+    async (req, res) => {
+        let sqlQuery = 'DELETE FROM tasks WHERE tasks.id=? AND userId=?';
+        const {id} = req.authData;
+        let rows = await pool.query(sqlQuery, [req.params.id, id]);
+        if (rows.affectedRows > 0) {
+            res.status(200).json(
+                {
+                    status: 'ok',
+                    deleted: true,
+                }
+            );
+        } else {
+            res.status(404).json({
+                status: 'error',
+                deleted: false,
+                message: `the task don´t exist or is not yours`,
+            });
+        }
     }
-}));
+)));
 
 module.exports = tasksRouter;
